Move product queries business logic to core-products package #275
