+++ packages/api/resolvers/queries/product.js
@@ -4,14 +4,8 @@ import { ProductNotFoundError, InvalidIdError } from '../../errors';
 
 export default function product(root, { productId, slug }, { userId }) {
   log(`query product ${productId} ${slug}`, { userId });
-
   if (!productId === !slug) throw new InvalidIdError({ productId, slug });
-
-  const foundProduct = productId
-    ? Products.findOne({ _id: productId })
-    : Products.findOne({ slugs: slug });
-
+  const foundProduct = Products.findProduct({ productId, slug });
   if (!foundProduct) throw new ProductNotFoundError({ productId });
-
   return foundProduct;
 }
